ansible modification and S3 env creation

+++ Kubernetes_v6/Addendum/Addenda13/all_in_one.sh
@@ -1,4 +1,5 @@
-mv /etc/ansible/hosts /etc/ansible/hosts.bak
+mkdir -p /etc/ansible
+if [ -f /etc/ansible/hosts ]; then mv /etc/ansible/hosts /etc/ansible/hosts.bak; fi;
 cp hosts /etc/ansible/ 
 
 ansible-playbook svm_secondary_create.yaml
