+++ Kubernetes_v6/Scenarios/Scenario24/setup_sc24_svm.sh
@@ -3,8 +3,8 @@ if [[ $(dnf list installed | grep ansible-core | wc -l) -eq 0 ]]; then
   echo "# Ansible install"
   echo "##############################################################"
   
-  dnf install -y ansible python-pip
-  pip install netapp-lib
+  dnf install -y python-pip
+  pip install ansible-core==2.15.12 netapp-lib
   ansible-galaxy collection install netapp.ontap --ignore-certs
 fi
 
@@ -13,7 +13,8 @@ echo "##############################################"
 echo "# Ansible Config"
 echo "##############################################"
 
-mv /etc/ansible/hosts /etc/ansible/hosts.bak
+mkdir -p /etc/ansible
+if [ -f /etc/ansible/hosts ]; then mv /etc/ansible/hosts /etc/ansible/hosts.bak; fi;
 cp ../../Addendum/Addenda13/hosts /etc/ansible/ 
 
 echo
